+++ tools/others/email_verifier.py
@@ -7,10 +7,12 @@ class KnockMail(HackingTool):
     TITLE = "Knockmail"
     DESCRIPTION = "KnockMail Tool Verify If Email Exists"
     INSTALL_COMMANDS = [
-        "git clone https://github.com/4w4k3/KnockMail.git",
-        "cd KnockMail;sudo pip install -r requeriments.txt"
+        dict(cmd="git clone https://github.com/4w4k3/KnockMail.git"),
+        dict(cmd="sudo pip install -r requeriments.txt", cwd="KnockMail")
     ]
-    RUN_COMMANDS = ["cd KnockMail;python knock.py"]
+    RUN_COMMANDS = [
+        dict(cmd="python knock.py", cwd="KnockMail"),
+    ]
     PROJECT_URL = "https://github.com/4w4k3/KnockMail"
